Refactor into tracing interfaces
