+++ inject/src/main/java/io/micronaut/context/BeanResolutionTraceConfiguration.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2017-2025 original authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.micronaut.context;
+
+import io.micronaut.core.annotation.Nullable;
+import java.util.Set;
+
+/**
+ * Configuration for tracing of bean resolution.
+ *
+ * @param mode          The mode
+ * @param classPatterns The class patterns to include
+ * @param destination   The destination to write data to if applicable.
+ */
+public record BeanResolutionTraceConfiguration(
+    BeanResolutionTraceMode mode,
+    Set<String> classPatterns,
+    @Nullable
+    String destination
+) {
+    public BeanResolutionTraceConfiguration {
+        if (classPatterns == null) {
+            classPatterns = BeanResolutionTraceMode.getDefaultTraceClasses();
+        }
+        if (mode == null) {
+            mode = BeanResolutionTraceMode.getDefaultMode(classPatterns);
+        }
+    }
+
+    public BeanResolutionTraceConfiguration() {
+        this(BeanResolutionTraceMode.NONE, Set.of(), null);
+    }
+}
